requirements.txt
appnope==0.1.0
astroid==2.0.4
backcall==0.1.0
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.5
django-braces==1.13.0
django-extensions==2.1.2
django-heroku==0.3.1
django-model-utils==3.1.2
django-star-ratings==0.7.0
google-api-core==1.4.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-vision==0.34.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
httplib2==0.11.3
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pylint==2.1.1
pytz==2018.5
requests==2.19.1
rsa==4.0
simplegeneric==0.8.1
six==1.11.0
swapper==1.1.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
whitenoise==4.1
wrapt==1.10.11
